transactions are working in userdao with annotation

+++ src/main/java/com/codecool/restauratio/services/UserService.java
@@ -40,7 +40,7 @@ public class UserService {
 
     // returns the id of the created restaurant pared with this user
     private int newRestaurant(String name, String description, String location, List<Food> menu, int capacity, int userId) throws ConnectToDBFailed, FailedDataVertification {
-        User owner = userDao.getById((long) userId);
+        User owner = userDao.getById(userId);
         Restaurant restaurant = new Restaurant(name, description, location, menu, capacity, owner);
         RestaurantDao restaurantDao = new RestaurantDao();
         restaurantDao.add(restaurant);
